requirements.txt
# Core dependencies
openai>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Configuration management
PyYAML>=6.0

# Agent framework - OpenAI Agents SDK
openai-agents>=0.1.0

# Async HTTP client for API calls
aiohttp>=3.8.0

# For Kubernetes integration
# kubectl should be installed separately and configured
# No Python package needed - uses subprocess

# For HTTP requests and validation
requests>=2.28.0

# Web framework
tornado>=6.0.0

# For data handling
python-dateutil>=2.8.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

